Revert "[RFR][1LP] move conftest into pytest plugin"
